+++ src/projectit.cpp
@@ -161,12 +161,12 @@ SEXP transform_ng(SEXP fromargs, SEXP toargs, SEXP coordOp, SEXP npts, SEXP x, S
     return(R_NilValue);
 }
 
-SEXP project_ng(SEXP n, SEXP xlon, SEXP ylat, SEXP zz, SEXP inv, SEXP ob_tran, SEXP coordOp) {
+SEXP project_ng(SEXP n, SEXP xlon, SEXP ylat, SEXP zz, SEXP coordOp) {
     Rprintf("Not available for deprecated API");
     return(R_NilValue);
 }
 
-SEXP project_ng_coordOp(SEXP proj, SEXP inv, SEXP aoi) {
+SEXP project_ng_coordOp(SEXP proj, SEXP inv, SEXP aoi, SEXP ob_tran) {
     Rprintf("Not available for deprecated API");
     return(R_NilValue);
 }